build: convert release scripts to typescript
