+++ package.json
@@ -22,7 +22,7 @@
     "typecheck:packages": "pnpm -r --filter \"@daypicker/*\" typecheck && pnpm --filter react-day-picker typecheck",
     "check:versions": "node ./scripts/check-package-versions.mjs",
     "pack:dry-run": "node ./scripts/npm-pack-dry-run.mjs",
-    "release:ci": "node ./scripts/release-ci.mjs",
+    "release:ci": "pnpm exec tsx ./scripts/release-ci.ts",
     "typecheck-watch": "tsc --project ./tsconfig.json --noEmit --watch"
   },
   "devDependencies": {
@@ -53,6 +53,7 @@
     "react-day-picker": "workspace:*",
     "react-dom": "^19.2.5",
     "ts-node": "^10.9.2",
+    "tsx": "^4.20.6",
     "tslib": "^2.8.1",
     "typescript": "^6.0.3",
     "typescript-css-modules": "^1.0.4"
